Update to 25.05 and fix GDM
For the 25.05 upgrade I had to nuke some of my personal modules because
I see that nix people have been busy with making them work and exist
within home-manager, thank you for that.

GDM had a stupid issue where the session would not initialize the XDG
paths properly because it would launch the native system's sway binary
instead of the one provided by home-manager and that would result in
apps breaking to the PATHs not being set up properly. This "fix" just
removes the native sway binary and adds a wayland-sessions file that
launches sway through the fish shell instead which would initialize
all of the environment variables properly because fish has
initialization scripts set up so that the Nix stuff works fine with it.

...yay

...and an another bunch of bullshit that just me using the system and
installing more and more crap as time goes on.